+++ apps/kimi-code/src/tui/controllers/config-commands.ts
@@ -1,0 +1,383 @@
+import type { PermissionMode, Session } from '@moonshot-ai/kimi-code-sdk';
+
+import { EditorSelectorComponent } from '../components/dialogs/editor-selector';
+import { ModelSelectorComponent } from '../components/dialogs/model-selector';
+import { PermissionSelectorComponent } from '../components/dialogs/permission-selector';
+import { SettingsSelectorComponent, type SettingsSelection } from '../components/dialogs/settings-selector';
+import { ThemeSelectorComponent } from '../components/dialogs/theme-selector';
+import { saveTuiConfig } from '../config';
+import type { Theme } from '../theme';
+import { NO_ACTIVE_SESSION_MESSAGE } from '../constant/kimi-tui';
+import { isTheme } from '../theme/index';
+import { formatErrorMessage } from '../utils/event-payload';
+import { showUsage } from './info-commands';
+import type { SlashCommandHost } from './slash-commands';
+
+// ---------------------------------------------------------------------------
+// Plan / Config commands
+// ---------------------------------------------------------------------------
+
+export async function handlePlanCommand(host: SlashCommandHost, args: string): Promise<void> {
+  const session = host.session;
+  if (session === undefined) {
+    host.showError(NO_ACTIVE_SESSION_MESSAGE);
+    return;
+  }
+
+  const subcmd = args.trim().toLowerCase();
+  if (subcmd === 'clear') {
+    await session.clearPlan();
+    host.showNotice('Plan cleared');
+    return;
+  }
+
+  let enabled: boolean;
+  if (subcmd.length === 0) enabled = !host.state.appState.planMode;
+  else if (subcmd === 'on') enabled = true;
+  else if (subcmd === 'off') enabled = false;
+  else {
+    host.showError(`Unknown plan subcommand: ${subcmd}`);
+    return;
+  }
+
+  await applyPlanMode(host, session, enabled);
+}
+
+async function applyPlanMode(host: SlashCommandHost, session: Session, enabled: boolean): Promise<void> {
+  try {
+    await session.setPlanMode(enabled);
+    host.setAppState({ planMode: enabled });
+    if (enabled) {
+      const plan = await session.getPlan().catch(() => null);
+      host.showNotice(
+        'Plan mode: ON',
+        plan?.path !== undefined ? `Plan will be created here: ${plan.path}` : undefined,
+      );
+      return;
+    }
+    host.showNotice('Plan mode: OFF');
+  } catch (error) {
+    const msg = formatErrorMessage(error);
+    host.showError(`Failed to set plan mode: ${msg}`);
+  }
+}
+
+export async function handleYoloCommand(host: SlashCommandHost, args: string): Promise<void> {
+  const session = host.session;
+  if (session === undefined) {
+    host.showError(NO_ACTIVE_SESSION_MESSAGE);
+    return;
+  }
+
+  let enabled: boolean;
+  if (args === 'on') enabled = true;
+  else if (args === 'off') enabled = false;
+  else enabled = host.state.appState.permissionMode !== 'yolo';
+
+  await session.setPermission(enabled ? 'yolo' : 'manual');
+  host.setAppState({ permissionMode: enabled ? 'yolo' : 'manual' });
+  if (enabled) {
+    host.showNotice(
+      'YOLO mode: ON',
+      'All actions will be approved automatically. Use with caution.',
+    );
+    return;
+  }
+  host.showNotice('YOLO mode: OFF');
+}
+
+export async function handleCompactCommand(host: SlashCommandHost, args: string): Promise<void> {
+  const session = host.session;
+  if (session === undefined) {
+    host.showError(NO_ACTIVE_SESSION_MESSAGE);
+    return;
+  }
+  const customInstruction = args.trim() || undefined;
+  await session.compact({ instruction: customInstruction });
+}
+
+export async function handleEditorCommand(host: SlashCommandHost, args: string): Promise<void> {
+  const command = args.trim();
+  if (command.length === 0) {
+    showEditorPicker(host);
+    return;
+  }
+  await applyEditorChoice(host, command);
+}
+
+export async function handleThemeCommand(host: SlashCommandHost, args: string): Promise<void> {
+  const theme = args.trim();
+  if (theme.length === 0) {
+    showThemePicker(host);
+    return;
+  }
+  if (!isTheme(theme)) {
+    host.showError(`Unknown theme: ${theme}`);
+    return;
+  }
+  await applyThemeChoice(host, theme);
+}
+
+export function handleModelCommand(host: SlashCommandHost, args: string): void {
+  const alias = args.trim();
+  if (alias.length === 0) {
+    showModelPicker(host);
+    return;
+  }
+  if (host.state.appState.availableModels[alias] === undefined) {
+    host.showError(`Unknown model alias: ${alias}`);
+    return;
+  }
+  showModelPicker(host, alias);
+}
+
+// ---------------------------------------------------------------------------
+// Pickers & config apply
+// ---------------------------------------------------------------------------
+
+function showEditorPicker(host: SlashCommandHost): void {
+  const currentValue = host.state.appState.editorCommand ?? '';
+  host.mountEditorReplacement(
+    new EditorSelectorComponent({
+      currentValue,
+      colors: host.state.theme.colors,
+      onSelect: (value) => {
+        host.restoreEditor();
+        void applyEditorChoice(host, value);
+      },
+      onCancel: () => {
+        host.restoreEditor();
+      },
+    }),
+  );
+}
+
+async function applyEditorChoice(host: SlashCommandHost, value: string): Promise<void> {
+  const previous = host.state.appState.editorCommand ?? '';
+  if (value === previous && value.length > 0) {
+    host.showStatus(`Editor unchanged: ${value.length > 0 ? value : 'auto-detect'}`);
+    return;
+  }
+
+  const editorCommand = value.length > 0 ? value : null;
+  try {
+    await saveTuiConfig({
+      theme: host.state.appState.theme,
+      editorCommand,
+      notifications: host.state.appState.notifications,
+    });
+  } catch (error) {
+    host.showStatus(
+      `Failed to save editor: ${formatErrorMessage(error)}`,
+      host.state.theme.colors.error,
+    );
+    return;
+  }
+
+  host.setAppState({ editorCommand });
+  host.showStatus(
+    value.length > 0
+      ? `Editor set to "${value}".`
+      : 'Editor set to auto-detect ($VISUAL / $EDITOR).',
+  );
+}
+
+export function showModelPicker(host: SlashCommandHost, selectedValue: string = host.state.appState.model): void {
+  const entries = Object.entries(host.state.appState.availableModels);
+  if (entries.length === 0) {
+    host.showNotice(
+      'No models configured',
+      'Run /login to sign in to Kimi, or /connect to add another provider from a model catalog.',
+    );
+    return;
+  }
+  host.mountEditorReplacement(
+    new ModelSelectorComponent({
+      models: host.state.appState.availableModels,
+      currentValue: host.state.appState.model,
+      selectedValue,
+      currentThinking: host.state.appState.thinking,
+      colors: host.state.theme.colors,
+      searchable: true,
+      onSelect: ({ alias, thinking }) => {
+        host.restoreEditor();
+        void performModelSwitch(host, alias, thinking);
+      },
+      onCancel: () => {
+        host.restoreEditor();
+      },
+    }),
+  );
+}
+
+async function performModelSwitch(host: SlashCommandHost, alias: string, thinking: boolean): Promise<void> {
+  if (host.state.appState.streamingPhase !== 'idle') {
+    host.showError('Cannot switch models while streaming — press Esc or Ctrl-C first.');
+    return;
+  }
+
+  const level = thinking ? 'on' : 'off';
+  const prevModel = host.state.appState.model;
+  const prevThinking = host.state.appState.thinking;
+  const runtimeChanged = alias !== prevModel || thinking !== prevThinking;
+
+  const session = host.session;
+  try {
+    if (session === undefined && runtimeChanged) {
+      await host.authFlow.activateModelAfterLogin(alias, thinking);
+    } else if (session !== undefined) {
+      if (alias !== prevModel) {
+        await session.setModel(alias);
+      }
+      if (thinking !== prevThinking) {
+        await session.setThinking(level);
+      }
+    }
+  } catch (error) {
+    const msg = formatErrorMessage(error);
+    host.showError(`Failed to switch model: ${msg}`);
+    return;
+  }
+
+  host.setAppState({ model: alias, thinking });
+  if (session === undefined && runtimeChanged) {
+    if (alias !== prevModel) {
+      host.track('model_switch', { model: alias });
+    }
+    if (thinking !== prevThinking) {
+      host.track('thinking_toggle', { enabled: thinking });
+    }
+  }
+
+  let persisted = false;
+  try {
+    persisted = await persistModelSelection(host, alias, thinking);
+  } catch (error) {
+    const msg = formatErrorMessage(error);
+    host.showError(`Switched to ${alias}, but failed to save default: ${msg}`);
+    return;
+  }
+
+  const status = runtimeChanged
+    ? `Switched to ${alias} with thinking ${level}.`
+    : persisted
+      ? `Saved ${alias} with thinking ${level} as default.`
+      : `Already using ${alias} with thinking ${level}.`;
+  host.showStatus(status, host.state.theme.colors.success);
+}
+
+async function persistModelSelection(host: SlashCommandHost, alias: string, thinking: boolean): Promise<boolean> {
+  const config = await host.harness.getConfig({ reload: true });
+  if (config.defaultModel === alias && config.defaultThinking === thinking) {
+    return false;
+  }
+  await host.harness.setConfig({
+    defaultModel: alias,
+    defaultThinking: thinking,
+  });
+  return true;
+}
+
+function showThemePicker(host: SlashCommandHost): void {
+  host.mountEditorReplacement(
+    new ThemeSelectorComponent({
+      currentValue: host.state.appState.theme,
+      colors: host.state.theme.colors,
+      onSelect: (value) => {
+        host.restoreEditor();
+        void applyThemeChoice(host, value);
+      },
+      onCancel: () => {
+        host.restoreEditor();
+      },
+    }),
+  );
+}
+
+async function applyThemeChoice(host: SlashCommandHost, theme: Theme): Promise<void> {
+  if (theme === host.state.appState.theme) {
+    if (theme === 'auto') host.refreshTerminalThemeTracking();
+    host.showStatus(`Theme unchanged: "${theme}".`);
+    return;
+  }
+
+  try {
+    await saveTuiConfig({
+      theme,
+      editorCommand: host.state.appState.editorCommand,
+      notifications: host.state.appState.notifications,
+    });
+  } catch (error) {
+    host.showStatus(
+      `Failed to save theme: ${formatErrorMessage(error)}`,
+      host.state.theme.colors.error,
+    );
+    return;
+  }
+
+  const resolved = theme === 'auto' ? host.state.theme.resolvedTheme : theme;
+  host.applyTheme(theme, resolved);
+  host.refreshTerminalThemeTracking();
+  host.track('theme_switch', { theme });
+  const detail = theme === 'auto' ? ` (tracking terminal; current: ${resolved})` : '';
+  host.showStatus(`Theme set to "${theme}"${detail}.`);
+}
+
+export function showPermissionPicker(host: SlashCommandHost): void {
+  host.mountEditorReplacement(
+    new PermissionSelectorComponent({
+      currentValue: host.state.appState.permissionMode,
+      colors: host.state.theme.colors,
+      onSelect: (value) => {
+        host.restoreEditor();
+        void applyPermissionChoice(host, value);
+      },
+      onCancel: () => {
+        host.restoreEditor();
+      },
+    }),
+  );
+}
+
+async function applyPermissionChoice(host: SlashCommandHost, mode: PermissionMode): Promise<void> {
+  if (mode === host.state.appState.permissionMode) {
+    host.showStatus(`Permission mode unchanged: ${mode}.`);
+    return;
+  }
+
+  try {
+    await host.requireSession().setPermission(mode);
+  } catch (error) {
+    const msg = formatErrorMessage(error);
+    host.showError(`Failed to set permission mode: ${msg}`);
+    return;
+  }
+
+  host.setAppState({ permissionMode: mode });
+  host.showNotice(`Permission mode: ${mode}`);
+}
+
+export function showSettingsSelector(host: SlashCommandHost): void {
+  host.mountEditorReplacement(
+    new SettingsSelectorComponent({
+      colors: host.state.theme.colors,
+      onSelect: (value) => {
+        handleSettingsSelection(host, value);
+      },
+      onCancel: () => {
+        host.restoreEditor();
+      },
+    }),
+  );
+}
+
+function handleSettingsSelection(host: SlashCommandHost, value: SettingsSelection): void {
+  host.restoreEditor();
+  switch (value) {
+    case 'model': showModelPicker(host); return;
+    case 'permission': showPermissionPicker(host); return;
+    case 'theme': showThemePicker(host); return;
+    case 'editor': showEditorPicker(host); return;
+    case 'usage': void showUsage(host); return;
+  }
+}
