+++ apps/kimi-code/src/tui/controllers/info-commands.ts
@@ -1,0 +1,185 @@
+import { release as osRelease, type as osType } from 'node:os';
+
+import type { McpServerInfo, SessionStatus, SessionUsage } from '@moonshot-ai/kimi-code-sdk';
+
+import { buildMcpStatusReportLines } from '../components/messages/mcp-status-panel';
+import { buildStatusReportLines } from '../components/messages/status-panel';
+import { buildUsageReportLines, UsagePanelComponent, type ManagedUsageReport } from '../components/messages/usage-panel';
+import {
+  FEEDBACK_ISSUE_URL,
+  FEEDBACK_STATUS_CANCELLED,
+  FEEDBACK_STATUS_FALLBACK,
+  FEEDBACK_STATUS_NOT_SIGNED_IN,
+  FEEDBACK_STATUS_SUBMITTING,
+  FEEDBACK_STATUS_SUCCESS,
+  FEEDBACK_TELEMETRY_EVENT,
+  feedbackSessionLine,
+  withFeedbackVersionPrefix,
+} from '../constant/feedback';
+import { isManagedUsageProvider } from '../constant/kimi-tui';
+import { formatErrorMessage } from '../utils/event-payload';
+import { openUrl } from '../utils/open-url';
+import { promptFeedbackInput } from './slash-command-prompts';
+import type { SlashCommandHost } from './slash-commands';
+
+// ---------------------------------------------------------------------------
+// Feedback
+// ---------------------------------------------------------------------------
+
+export async function handleFeedbackCommand(host: SlashCommandHost): Promise<void> {
+  const fallback = (reason: string): void => {
+    host.showStatus(reason);
+    host.showStatus(FEEDBACK_ISSUE_URL);
+    openUrl(FEEDBACK_ISSUE_URL);
+  };
+
+  const providerKey = host.state.appState.availableModels[host.state.appState.model]?.provider;
+  if (!isManagedUsageProvider(providerKey)) {
+    fallback(FEEDBACK_STATUS_NOT_SIGNED_IN);
+    return;
+  }
+
+  const content = await promptFeedbackInput(host);
+  if (content === undefined) {
+    host.showStatus(FEEDBACK_STATUS_CANCELLED);
+    return;
+  }
+
+  const spinner = host.showLoginProgressSpinner(FEEDBACK_STATUS_SUBMITTING);
+  const res = await host.harness.auth.submitFeedback({
+    content,
+    sessionId: host.state.appState.sessionId,
+    version: withFeedbackVersionPrefix(host.state.appState.version),
+    os: `${osType()} ${osRelease()}`,
+    model: host.state.appState.model.length > 0 ? host.state.appState.model : null,
+  });
+
+  if (res.kind === 'ok') {
+    spinner.stop({ ok: true, label: FEEDBACK_STATUS_SUCCESS });
+    host.showStatus(feedbackSessionLine(host.state.appState.sessionId));
+    host.track(FEEDBACK_TELEMETRY_EVENT);
+    return;
+  }
+
+  spinner.stop({ ok: false, label: res.message });
+  fallback(FEEDBACK_STATUS_FALLBACK);
+}
+
+// ---------------------------------------------------------------------------
+// Info commands
+// ---------------------------------------------------------------------------
+
+interface SessionUsageResult {
+  readonly usage?: SessionUsage;
+  readonly error?: string;
+}
+
+interface RuntimeStatusResult {
+  readonly status?: SessionStatus;
+  readonly error?: string;
+}
+
+interface ManagedUsageResult {
+  readonly usage?: ManagedUsageReport;
+  readonly error?: string;
+}
+
+export async function showUsage(host: SlashCommandHost): Promise<void> {
+  const sessionUsage = await loadSessionUsageReport(host);
+  const managedUsage = await loadManagedUsageReport(host);
+  const lines = buildUsageReportLines({
+    colors: host.state.theme.colors,
+    sessionUsage: sessionUsage.usage,
+    sessionUsageError: sessionUsage.error,
+    contextUsage: host.state.appState.contextUsage,
+    contextTokens: host.state.appState.contextTokens,
+    maxContextTokens: host.state.appState.maxContextTokens,
+    managedUsage: managedUsage?.usage,
+    managedUsageError: managedUsage?.error,
+  });
+  const panel = new UsagePanelComponent(lines, host.state.theme.colors.primary);
+  host.state.transcriptContainer.addChild(panel);
+  host.state.ui.requestRender();
+}
+
+export async function showStatusReport(host: SlashCommandHost): Promise<void> {
+  const [runtimeStatus, managedUsage] = await Promise.all([
+    loadRuntimeStatusReport(host),
+    loadManagedUsageReport(host),
+  ]);
+  const appState = host.state.appState;
+  const lines = buildStatusReportLines({
+    colors: host.state.theme.colors,
+    version: appState.version,
+    model: appState.model,
+    workDir: appState.workDir,
+    sessionId: appState.sessionId,
+    sessionTitle: appState.sessionTitle,
+    thinking: appState.thinking,
+    permissionMode: appState.permissionMode,
+    planMode: appState.planMode,
+    contextUsage: appState.contextUsage,
+    contextTokens: appState.contextTokens,
+    maxContextTokens: appState.maxContextTokens,
+    availableModels: appState.availableModels,
+    status: runtimeStatus.status,
+    statusError: runtimeStatus.error,
+    managedUsage: managedUsage?.usage,
+    managedUsageError: managedUsage?.error,
+  });
+  const panel = new UsagePanelComponent(lines, host.state.theme.colors.primary, ' Status ');
+  host.state.transcriptContainer.addChild(panel);
+  host.state.ui.requestRender();
+}
+
+export async function showMcpServers(host: SlashCommandHost): Promise<void> {
+  let servers: readonly McpServerInfo[];
+  try {
+    servers = await host.requireSession().listMcpServers();
+  } catch (error) {
+    host.showError(`Failed to load MCP servers: ${formatErrorMessage(error)}`);
+    return;
+  }
+
+  const lines = buildMcpStatusReportLines({
+    colors: host.state.theme.colors,
+    servers,
+  });
+  const title = servers.length > 0 ? ` MCP (${servers.length}) ` : ' MCP ';
+  const panel = new UsagePanelComponent(lines, host.state.theme.colors.primary, title);
+  host.state.transcriptContainer.addChild(panel);
+  host.state.ui.requestRender();
+}
+
+async function loadSessionUsageReport(host: SlashCommandHost): Promise<SessionUsageResult> {
+  try {
+    return { usage: await host.requireSession().getUsage() };
+  } catch (error) {
+    return { error: formatErrorMessage(error) };
+  }
+}
+
+async function loadRuntimeStatusReport(host: SlashCommandHost): Promise<RuntimeStatusResult> {
+  try {
+    return { status: await host.requireSession().getStatus() };
+  } catch (error) {
+    return { error: error instanceof Error ? error.message : String(error) };
+  }
+}
+
+async function loadManagedUsageReport(host: SlashCommandHost): Promise<ManagedUsageResult | undefined> {
+  const alias = host.state.appState.model;
+  const providerKey = host.state.appState.availableModels[alias]?.provider;
+  if (!isManagedUsageProvider(providerKey)) return undefined;
+
+  let res;
+  try {
+    res = await host.harness.auth.getManagedUsage(providerKey);
+  } catch (error) {
+    return { error: formatErrorMessage(error) };
+  }
+  if (res.kind === 'error') {
+    return { error: res.message };
+  }
+  return { usage: { summary: res.summary, limits: res.limits } };
+}
